python/pdstools/utils/anonymizer/requirements.txt
python==3.10.8

numpy==1.24.1
fire==0.5.0
pyyaml==6.0